Change cancel link on usps to recognize pending user
**Why**: A user that has a pending profile can choose to receive another
letter on the verify account screen. That takes them to the USPS page.
When they click cancel there, they are not cancelling IdV and are
instead cancelling the action to resend their letter. This commit
changes the cancel link so that in this case it takes them back to the
verify account step and not to the idv cancellation screen.

This commit also rename the UspsDecorator to be a UspsPresenter since
it doesn't make sense as a decorator.